Removed unused code. Implemented Disconnect Method. Added Label to Allow Untrusted Certificate switch. Moved Some Classes in extra files. Implemented resolving of typeId if it is not present in the BrowseResult.
