+++ umatiGateway/Pages/PubSub.cshtml
@@ -24,6 +24,7 @@
             <input type="text" class="settings-input" id="Prefix" name="MqttPrefix" value="@Model.UmatiGatewayApp.ActiveConfiguration.PubSubProviderConfig.Prefix">
             <label for="MqttConnection" class="settings-label">ConnectionUrl</label>
             <input id="MqttConnection" class="settings-input" type="text" name="ConnectionUrl" value="@Model.UmatiGatewayApp.ActiveConfiguration.PubSubProviderConfig.ServerEndpoint">
+            <label for="PubSubAllowUntrustedCertificates" class="settings-label">Allow Untrusted Certificates</label>
             <div class="settings-on-off-label">
                 <span class="off">Off</span>
                 <label class="settings-switch">
